feat: update cart

+++ src/models/Product.js
@@ -20,7 +20,7 @@ class Product extends Model {
                     allowNull: false,
                 },
                 price: {
-                    type: Sequelize.INTEGER,
+                    type: Sequelize.DECIMAL(10, 2),
                     allowNull: false,
                 },
                 description: {
